+++ baiducloud-python-sdk-bls/sample/example_delete_alarm_policy.py
@@ -18,7 +18,7 @@ if __name__ == '__main__':
             credentials=BceCredentials(access_key_id, secret_access_key), endpoint=endpoint
         )
         client = BlsClient(bce_client_config)
-        request = bls_models.DeleteAlarmPolicyRequest(name=[])
+        request = bls_models.DeleteAlarmPolicyRequest(names=[])
         res = client.delete_alarm_policy(request)
         print(res.to_json_string())
     except exception.BceHttpClientError as e: